requirements.txt
# Core dependencies
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
tensorflow==2.16.1
PyYAML>=6.0
joblib==1.4.0
openpyxl>=3.0.0

# Development tools
black==24.1.1
ruff==0.2.1
pytest==8.0.0
pre-commit==3.6.0

# Optional: Experiment tracking
mlflow>=2.0.0

# Optional: Visualization
matplotlib>=3.4.0
seaborn>=0.11.0

# Optional: Jupyter for notebooks
jupyter==1.0.0
ipykernel==6.29.0

# Additional dependencies
torch>=2.0.0
tqdm>=4.62.0 